core-app-api: extract out overrideBaseUrlConfigs
